requirements.txt
torch>=2.0
torchvision
numpy
matplotlib
pandas
wandb
tqdm
einops
jupyter
python-dotenv
typer
rich
xai-sdk
